+++ pytest.ini
@@ -2,3 +2,4 @@
 markers =
     remote: Tests accessing the internet.
     slow: Slow tests.
+    sql: All tests related to SQL.
