+++ wetterdienst/dwd/__init__.py
@@ -1,0 +1,2 @@
+# Load Pandas DataFrame extension.
+import wetterdienst.dwd.pandas  # noqa:F401
