+++ wetterdienst/dwd/metadata/__init__.py
@@ -1,0 +1,3 @@
+from .parameter import Parameter  # noqa:F401
+from .period_type import PeriodType  # noqa:F401
+from .time_resolution import TimeResolution  # noqa:F401
