+++ wetterdienst/exceptions.py
@@ -1,6 +1,4 @@
 class FailedDownload(Exception):
-    """ Failed download exception """
-
     pass
