--- wetterdienst/pandas/__init__.py
@@ -1,2 +1,0 @@
-import wetterdienst.pandas.io  # noqa:F401
-import wetterdienst.pandas.wd  # noqa:F401
