+++ wetterdienst/util/__init__.py
@@ -1,0 +1,2 @@
+# Load Pandas DataFrame extension.
+import wetterdienst.util.pandas  # noqa:F401
